Added a clean target to the Makefile
git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@2625 fc592f45-0881-4044-91ff-c96495a54512

+++ www/build-www.py
@@ -127,6 +127,8 @@ def create_license_html(package_name, packdir, outdir):
 
 def rm_rf(p):
     #print("Deleting " + p);
+    if (p == "./CVS"):
+        return;
     for root, dirs, files in os.walk(p, topdown=False):
         for name in files:
             os.remove(os.path.join(root, name));
